fixed errors and warnings in bp_init

+++ bp.cpp
@@ -45,8 +45,8 @@ int BP_init(unsigned btbSize, unsigned historySize, unsigned tagSize, unsigned f
 	memset(table_ptr -> rows, 0, sizeof(BTB_row) * btbSize);
 
 	table_ptr->btbSize = btbSize;
-	int flag_Size = 0;
-	for(int i=1; i<=32; i=i*2){
+	unsigned flag_Size = 0;
+	for(unsigned i=1; i<=32; i=i*2){
 		if(btbSize == i){
 			flag_Size = 1;
 		}
@@ -57,7 +57,7 @@ int BP_init(unsigned btbSize, unsigned historySize, unsigned tagSize, unsigned f
 	if(historySize < 1 || historySize > 8) return -1;
 
 	table_ptr->tagSize = tagSize;
-	if((tagSize < 0) || (tagSize > 30 - static_cast<int>(log2(btbSize)) ) ) return -1;
+	if((tagSize < 0) || (tagSize > 30 - static_cast<unsigned>(log2(btbSize)) ) ) return -1;
 
 	table_ptr->fsmState = fsmState;
 	if( (fsmState < 0) || (fsmState > 3) ) return -1;
@@ -71,14 +71,14 @@ int BP_init(unsigned btbSize, unsigned historySize, unsigned tagSize, unsigned f
 
 		//access: *((state_chooser + yyyy) + history_reg
 		table_ptr->state_chooser = new int*[btbSize];
-		for(int i=0; i<btbSize; i++ ){
-			(table_ptr->state_chooser)[i] = new int[pow(2,historySize)];
+		for(unsigned i=0; i<btbSize; i++ ){
+			(table_ptr->state_chooser)[i] = new int[(int)pow(2,historySize)];
 			memset((table_ptr->state_chooser)[i], fsmState, (pow(2, historySize))*sizeof(int));
 		}
 	}
 	else{//global
 		table_ptr->state_chooser = new int*[1];
-		*(table_ptr->state_chooser) = new int[pow(2, historySize)-1];
+		*(table_ptr->state_chooser) = new int[(int)pow(2, historySize)-1];
 		memset(*(table_ptr->state_chooser), fsmState, (pow(2, historySize))*sizeof(int));
 	}
